requirements.txt
streamlit
PyMuPDF
python-docx
pandas
sentence-transformers
transformers
faiss-cpu
ctransformers